Fix failing integration tests

+++ SpareParts.API.Int.Tests/InventoryTests.cs
@@ -13,10 +13,14 @@ namespace SpareParts.API.Int.Tests
         {
             _testFixture = testFixture;
             _dataHelper = new DataHelper(_testFixture.DbContext);
-            
+
             // clear inventory tables between tests
-            _testFixture.DbContext.InventoryItems.RemoveRange(_testFixture.DbContext.InventoryItems);
-            _testFixture.DbContext.SaveChanges();
+            if (_testFixture.DbContext.InventoryItems.Any())
+            {
+                _testFixture.DbContext.InventoryItems.RemoveRange(_testFixture.DbContext.InventoryItems);
+                _testFixture.DbContext.Parts.RemoveRange(_testFixture.DbContext.Parts);
+                _testFixture.DbContext.SaveChanges();
+            }
             _testFixture.DbContext.ChangeTracker.Clear();
 
             // ensure client has Auth Header set
